+++ peers/src/main/scala/akka/wamp/client/TransportHandler.scala
@@ -6,14 +6,16 @@ import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.model.ws.{Message => WebSocketMessage, _}
 import akka.stream._
 import akka.stream.scaladsl._
-import akka.wamp.messages.{Message => WampMessage, _}
+import akka.wamp.messages.{Disconnected, Message => WampMessage, _}
 import akka.wamp.serialization._
 
 import scala.concurrent.Future
 
 
-private class TransportHandler extends Actor {
-
+private 
+class TransportHandler(clientRef: ActorRef) 
+  extends Actor with ActorLogging
+{
   /** The execution context */
   private implicit val ec = context.system.dispatcher
 
@@ -39,19 +41,19 @@ private class TransportHandler extends Actor {
   /** The serialization flows */
   // TODO [Provide wamp.2.msgpack subprotocol](https://github.com/angiolep/akka-wamp/issues/12)
   private val serializationFlows = new JsonSerializationFlows(validateStrictUris, disconnectOffendingPeers)
-  
+
   /** The client actor */
   private var client: ActorRef = _
 
   /** The outlet actor */
-  private var transport: ActorRef = _
-  
+  private var outletHandler: ActorRef = _
+
   /**
     * Handle CONNECT and DISCONNECT commands
     */
   override def receive: Receive = {
-    case cmd @ Connect(uri, subprotocol) => {
-      client = sender()
+    case cmd @ Connect(uri, subprotocol) =>
+      context become handleConnecting(cmd)
       try {
         val outgoingSource: Source[WampMessage, ActorRef] =
           Source.actorRef[WampMessage](0, OverflowStrategies.DropBuffer)
@@ -61,11 +63,9 @@ private class TransportHandler extends Actor {
             .webSocketClientFlow(WebSocketRequest(uri, subprotocol = Some(subprotocol)))
 
         val incomingSink: Sink[WampMessage, akka.NotUsed] =
-          Sink.actorRef[WampMessage](client, onCompleteMessage = Disconnected)
+          Sink.actorRef[WampMessage](self, onCompleteMessage = Disconnected)
 
-        if (subprotocol != "wamp.2.json") 
-          throw new IllegalArgumentException(s"$subprotocol is not supported") 
-        
+        // AKKA STREAM
         val (outgoingActor, upgradeResponse) =
           outgoingSource
             .via(serializationFlows.serialize)
@@ -73,44 +73,69 @@ private class TransportHandler extends Actor {
             .viaMat(serializationFlows.deserialize)(Keep.left)
             .toMat(incomingSink)(Keep.left)
             .run()
-        
-        // hold outlet reference for later usage and watch for its termination
-        transport = outgoingActor
-        context.watch(transport)
+
+        // hold outlet reference for later usage
+        outletHandler = outgoingActor
 
         // just like a regular http request we can get 404 NotFound etc.
         // that will be available from upgrade.response
         upgradeResponse.onSuccess { case upgrade =>
           if (upgrade.response.status == SwitchingProtocols) {
-            context become connected
-            client ! Connected(self)
-          } 
-          else {
-            client ! CommandFailed(cmd, new Exception(upgrade.response.toString))
+            context become handleConnected
+            clientRef ! Connected(self)
           }
         }
       } catch {
         case ex: Throwable =>
-          client ! CommandFailed(cmd, ex)
+          // NOTE:
+          // It happens when the client tries to connect
+          //    - to a malformed URL (e.g. "ws!malformed:9999/uri")
+          clientRef ! CommandFailed(cmd, ex)
       }
+  }
+
+  
+  def handleConnecting(cmd: Connect): Receive = {
+    case signal @ Status.Failure(ex) => {
+      // NOTE: 
+      // As documented for Sink.actorRef(), this signal is sent when  
+      // the above AKKA STREAM is completed with a failure, such as
+      //
+      //   - akka.stream.StreamTcpException: Tcp command [Connect(...)] failed
+      //   - java.lang.IllegalArgumentException: WebSocket upgrade did not finish because of 'unexpected status code: 404 Not Found'
+      //
+      clientRef ! CommandFailed(cmd, ex)
+      self ! PoisonPill
+      // TODO will it poison outletHandler too?
     }
   }
+
   
-  
-  
-  def connected: Receive = {
-    case msg: WampMessage =>
-      transport forward msg
-    
+  def handleConnected: Receive = {
+    case outgoing: WampMessage if sender() == clientRef =>
+      outletHandler forward outgoing
+      
+    case incoming: WampMessage =>
+      // NOTE:
+      // Incoming message are those received from router
+      // (e.g. WELCOME, ABORT, EVENT, ...)
+      clientRef forward incoming
+
     case cmd @ Disconnect =>
-      transport ! PoisonPill
-    
-    case Terminated(actor) =>
+      // NOTE:
+      // It happens when client send Disconnect to this handler
+      // so to successfully terminate the outletHandler
+      // RouterDisconnect will be sent as consequence
+      outletHandler ! PoisonPill
+      
+    case signal @ Disconnected =>
+      // As documented for Sink.actorRef(), this signal is sent when  
+      // the above AKKA STREAM is completed successful, such as 
+      // disconnection from router side
+      clientRef ! Disconnected
       self ! PoisonPill
-      client ! Disconnected
   }
 }
-
 
 
 /**
@@ -120,5 +145,5 @@ private[wamp] object TransportHandler {
   /**
     * Factory for [[TransportHandler]] instances
     */
-  def props() = Props(new TransportHandler())
+  def props(clientRef: ActorRef) = Props(new TransportHandler(clientRef))
 }