+++ peers/src/test/scala/akka/wamp/client/SessionSpec.scala
@@ -2,7 +2,7 @@ package akka.wamp.client
 
 import org.scalamock.scalatest.MockFactory
 
-class SessionSpec extends ClientFixtureSpec with MockFactory {
+class SessionSpec extends ClientBaseSpec with MockFactory {
 
   "A client session" should "close and reply GOODBYE upon receiving GOODBYE from router" in { f =>
     // TODO https://github.com/angiolep/akka-wamp/issues/11